refactor: organize components in AieraChat
