+++ src/modules/AieraChat/components/Messages/MessageFactory/Block/Chart/Area/index.tsx
