+++ src/modules/AieraChat/components/Messages/MessageFactory/Block/Chart/Bar/index.tsx
