+++ src/modules/AieraChat/components/Messages/MessageFactory/Block/Chart/Scatter/index.tsx
